feat(client): 컴포넌트 모듈화
