+++ apps/client/src/app/senior/[position]/page.tsx
@@ -1,9 +1,10 @@
 'use client';
 
-import { Box, Flex, Grid, Image, Link, Text } from '@chakra-ui/react';
+import { Box, Grid, Image, Link } from '@chakra-ui/react';
 import { useRouter } from 'next/navigation';
 import { Header } from '@/components/common';
 import Footer from '@/components/common/Footer';
+import SeniorCard from '@/components/SeniorCard';
 import type { Position } from '@/types';
 
 const POSITION_LIST: Position[] = [
@@ -27,6 +28,7 @@ const SeniorPage = ({ params }: SeniorProps) => {
       <Header />
       <Box width="100vw" backgroundColor="white">
         <Box margin="0 auto" paddingTop="48px" paddingBottom="64px" width="80%">
+          {/* 배너 */}
           <Image
             src="/assets/mock_banner.jpeg"
             marginBottom="48px"
@@ -50,61 +52,7 @@ const SeniorPage = ({ params }: SeniorProps) => {
             {Array(10)
               .fill('')
               .map(() => (
-                // 선배 카드
-                <Box
-                  transition="all 0.25s ease"
-                  padding="24px"
-                  border="1px solid"
-                  borderColor="gray.200"
-                  borderRadius="8px"
-                  _hover={{
-                    backgroundColor: 'gray.50',
-                    cursor: 'pointer',
-                  }}
-                  height="144px"
-                >
-                  <Flex gap="24px" height="100%">
-                    <Image
-                      src="/assets/mock_senior.jpeg"
-                      borderRadius="8px"
-                      width="95px"
-                      height="100%"
-                      alt="Senior"
-                    />
-                    <Flex flexDirection="column" justifyContent="center" height="100%">
-                      <Flex flexDirection="column" gap="6px">
-                        <Flex flexDirection="column">
-                          <Flex alignItems="center" gap="6px">
-                            <Text fontSize="24px" fontWeight="semibold">
-                              김석진
-                            </Text>
-                            <Text fontSize="16px" color="gray.900" fontWeight="medium">
-                              2기 - 토스페이먼츠
-                            </Text>
-                          </Flex>
-                          <Text maxWidth="100%" color="gray.600" fontSize="14px">
-                            즐거움을 토대로 토대를 만드는 도리"토스"를 좋아하는 개발자
-                          </Text>
-                        </Flex>
-                        <Box
-                          display="flex"
-                          alignItems="center"
-                          justifyContent="center"
-                          color="gray.700"
-                          border="1px solid"
-                          borderColor="gray.200"
-                          borderRadius="18px"
-                          fontSize="14px"
-                          fontWeight="medium"
-                          width="80px"
-                          height="24px"
-                        >
-                          프론트엔드
-                        </Box>
-                      </Flex>
-                    </Flex>
-                  </Flex>
-                </Box>
+                <SeniorCard />
               ))}
           </Grid>
         </Box>
