+++ apps/client/src/store/UserInformation/index.ts
@@ -1,2 +1,3 @@
 export { useUserInformation } from './useUserInformation';
 export { userInformationState } from './userInformationState';
+export { isLoginState } from './isLoginState';
